move: IIndicator & PlotDescription to separate file for general use & to prevent cicular imports
